Expose browser as a toolkit
